Refactor: move domain models and UI state to core modules
This commit refactors the project structure by moving common domain models and UI state definitions to core modules.

Key changes:

-   **`ArduinoDomain` and `StateObjectDomain`**: Moved from `domain/conectionInternet` to `domain/core/models`.
-   **`ArduinosState`**: Moved from `feature/conectionInternet` to `ui/core/models`.
-   **New `BluetoothDeviceDomain`**: Added to `domain/connectionBluetooth` to represent Bluetooth device information.
-   **New `ResponseState`**: Added to `domain/connectionBluetooth` as a sealed class to represent success or error states for responses.
-   Updated import paths in `ArduinoToArduinoDomainMapper` to reflect the new location of `ArduinoDomain` and `StateObjectDomain`.
