+++ domain/connectionBluetooth/src/commonMain/Kotlin/com/virtualworld/multiplatformiot.domain.connectionBluetooth/BluetoothDevice.kt
@@ -1,0 +1,7 @@
+package com.virtualworld.multiplatformiot.domain.connectionBluetooth
+
+data class BluetoothDeviceDomain(
+    val name: String,
+    val address: String,
+    val isConnected: Boolean = false
+)
