+++ domain/connectionBluetooth/src/commonMain/Kotlin/com/virtualworld/multiplatformiot.domain.connectionBluetooth/ResponseState.kt
@@ -1,0 +1,8 @@
+package com.virtualworld.multiplatformiot.domain.connectionBluetooth
+
+sealed class ResponseState<out T : Any> {
+
+    data class Success<out T : Any>(val result: T) : ResponseState<T>()
+
+    data class Error(val exception: Exception) : ResponseState<Nothing>()
+}
